Summary :-
Auto-Translate bot is a technology which helps to break the language barrier among many users. It’s single command have power to translate messages into multiple languages. The accuracy level of translated language should be accurate as per the input. This technology is used to advance the lifestyle and make is more flexible and easier. This project leads to improve the technical aspect in convenient way for the benefit of clients.

Link :-  https://github.com/JanhaviC15/MajorProjects/blob/master/AutoBotFinal.py